Major refactor to match repo style
