+++ src/hooks/useTableFilters.ts
@@ -142,3 +142,12 @@ export default function useTableFilters<T = Record<string, unknown>>({
     setFilterValues: setUrlFilterValues as (values: Partial<T>) => void,
   };
 }
+
+// LEGACY: For backwards compatibility, re-export useTableFilters with old signature.
+// New code should use useTableFilters directly.
+export function useFilters<T = Record<string, unknown>>(
+  params: Omit<FilterParams<T>, 'syncToUrl' | 'omitFromUrl'>
+): TableFilterType<T> {
+  const { filters } = useTableFilters<T>({ ...params, syncToUrl: false });
+  return filters;
+}
