+++ src/modules/audit/components/ClientAuditHistory.tsx
@@ -3,12 +3,12 @@ import { ContextualCollapsibleListsProvider } from '@/components/elements/Collap
 import { ColumnDef } from '@/components/elements/table/types';
 import PageTitle from '@/components/layout/PageTitle';
 import useSafeParams from '@/hooks/useSafeParams';
+import { useFilters } from '@/hooks/useTableFilters';
 import {
   AUDIT_HISTORY_COLUMNS,
   AUDIT_HISTORY_USER_COLUMNS,
 } from '@/modules/audit/components/auditHistoryColumnDefs';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   ClientAuditEventFilterOptions,
   GetClientAuditEventsDocument,
