+++ src/modules/ce/components/admin/CeClientEligibleUnitGroupsTable.tsx
@@ -2,10 +2,10 @@ import { Chip, Paper } from '@mui/material';
 import pluralize from 'pluralize';
 import React from 'react';
 
+import { useFilters } from '@/hooks/useTableFilters';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import { DataColumnDef } from '@/modules/dataFetching/types';
 import ProjectTypeChip from '@/modules/hmis/components/ProjectTypeChip';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import { ProjectDashboardRoutes } from '@/routes/routes';
 import {
   CeEligibleUnitGroupFieldsFragment,
