+++ src/modules/ce/components/directReferral/ProjectOutgoingReferralsTable.tsx
@@ -2,10 +2,10 @@ import { Paper } from '@mui/material';
 import React from 'react';
 import DateWithRelativeTooltip from '@/components/elements/DateWithRelativeTooltip';
 import { getViewClientMenuItem } from '@/components/elements/table/tableRowActionUtil';
+import { useFilters } from '@/hooks/useTableFilters';
 import ReferralStatusChip from '@/modules/ce/components/referral/ReferralStatusChip';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import { DataColumnDef } from '@/modules/dataFetching/types';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   EnrollmentDashboardRoutes,
   ProjectDashboardRoutes,
