+++ src/modules/ce/components/project/ProjectReferralsTable.tsx
@@ -1,9 +1,9 @@
 import { Paper } from '@mui/material';
 import React from 'react';
 import { ColumnDef } from '@/components/elements/table/types';
+import { useFilters } from '@/hooks/useTableFilters';
 import { REFERRAL_COLUMNS } from '@/modules/ce/referralColumns';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import { ProjectDashboardRoutes } from '@/routes/routes';
 import {
   CeReferralTableFieldsFragment,
