+++ src/modules/ce/components/unit/PrioritizedClientsTable.tsx
@@ -6,6 +6,7 @@ import TableRowActions from '@/components/elements/table/TableRowActions';
 import { BASE_ACTION_COLUMN_DEF } from '@/components/elements/table/tableRowActionUtil';
 import { ColumnDef } from '@/components/elements/table/types';
 import useDebouncedState from '@/hooks/useDebouncedState';
+import { useFilters } from '@/hooks/useTableFilters';
 import { configurableCeColumns } from '@/modules/ce/components/admin/AdminCeClientsTable';
 import StartReferralButton from '@/modules/ce/components/unit/StartReferralButton';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
@@ -14,7 +15,6 @@ import {
   parseAndFormatDateTime,
   parseHmisDateString,
 } from '@/modules/hmis/hmisUtil';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import { useProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
 import CommonSearchInput from '@/modules/search/components/CommonSearchInput';
 import {
