+++ src/modules/enrollment/components/EnrollmentAssessmentsTable.tsx
@@ -1,4 +1,5 @@
 import { ColumnDef } from '@/components/elements/table/types';
+import { useFilters } from '@/hooks/useTableFilters';
 import {
   ASSESSMENT_COLUMNS,
   ASSESSMENT_DETAILS_COL,
@@ -6,7 +7,6 @@ import {
 } from '@/modules/assessments/util';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import { assessmentDescription } from '@/modules/hmis/hmisUtil';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   AssessmentFieldsFragment,
   GetEnrollmentAssessmentsDocument,
