+++ src/modules/enrollment/components/HouseholdAssessmentsTable.tsx
@@ -1,4 +1,5 @@
 import { ColumnDef } from '@/components/elements/table/types';
+import { useFilters } from '@/hooks/useTableFilters';
 import {
   ASSESSMENT_CLIENT_NAME_COL,
   ASSESSMENT_COLUMNS,
@@ -10,7 +11,6 @@ import {
   assessmentDescription,
   clientBriefName,
 } from '@/modules/hmis/hmisUtil';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import {
   GetHouseholdAssessmentsDocument,
   GetHouseholdAssessmentsQuery,
