+++ src/modules/household/components/ManageHousehold.tsx
@@ -18,9 +18,9 @@ import { ColumnDef } from '@/components/elements/table/types';
 import NotFound from '@/components/pages/NotFound';
 import { useGlobalFeatureFlags } from '@/hooks/useGlobalFeatureFlags';
 import { useIsMobile } from '@/hooks/useIsMobile';
+import { useFilters } from '@/hooks/useTableFilters';
 import { SsnDobShowContextProvider } from '@/modules/client/providers/ClientSsnDobVisibility';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 import AssociatedHouseholdMembers from '@/modules/household/components/AssociatedHouseholdMembers';
 import HouseholdActionButtons from '@/modules/household/components/elements/HouseholdActionButtons';
 import { RecentHouseholdMember } from '@/modules/household/types';
