+++ src/modules/units/components/UnitManagementTable.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useMemo } from 'react';
 import { ColumnDef } from '@/components/elements/table/types';
+import { useFilters } from '@/hooks/useTableFilters';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { useFilters } from '@/modules/hmis/tableFilterUtil';
 
 import { useProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
