+++ src/utils/tableFilterUtil.ts
@@ -5,7 +5,12 @@ import { getSchemaForInputType } from '@/modules/hmis/hmisUtil';
 import { HmisEnums } from '@/types/gqlEnums';
 import { GqlInputObjectSchemaType } from '@/types/gqlObjects';
 import { PickListType } from '@/types/gqlTypes';
-import { FilterType, BaseFilter } from '@/types/tableFilterTypes';
+import {
+  FilterType,
+  BaseFilter,
+  TableFilterType,
+} from '@/types/tableFilterTypes';
+import { ensureArray } from '@/utils/arrays';
 
 // Default remote pick lists to use for filters, based on filter name
 const FILTER_NAME_TO_PICK_LIST = {
@@ -97,4 +102,44 @@ export const getFilter = (
   if (!fieldSchema) return null;
 
   return getFilterForType(fieldName, fieldSchema.type, filterPickListArgs);
+};
+
+/**
+ * Method for transforming filter values by pulling out dynamic filters into separate key.
+ * This is necessary when filtering by dynamic filters, which are backed by HMIS Table Configuration.
+ *
+ * Example transformation:
+ *
+ * { "ActualFilter": "1", "SomeDynamicFilter": "2"}
+ *     =>
+ * { "ActualFilter":  "1", "dynamicFilters": [ { "key": "SomeDynamicFilter", "values": ["2"] }] }
+ */
+export const transformDynamicFilters = <FilterOptionsType>(
+  filters?: TableFilterType<FilterOptionsType>, // Filter configuration, so we know which ones are dynamic
+  filterValues?: Partial<FilterOptionsType> // Current filter values
+) => {
+  if (!filterValues) return; // No filter values, nothing to transform
+
+  // `filterValues` may be present even if `filters` is absent, on tables where `defaultFilterValues` are used without any user-facing filters (such as Bulk Services).
+  // In this case, we know the filters aren't dynamic so we don't need to transform them, just return `filterValues` as-is
+  if (!filters) return filterValues;
+
+  // Split filter values into static and dynamic
+  const staticFilterValues: Record<string, unknown> = {};
+  const dynamicFilters: Array<{ key: string; values: any[] }> = []; // { dynamic_key => [value1, value2, ...] }
+
+  for (const [key, value] of Object.entries(filterValues)) {
+    const isDynamic = filters[key as keyof FilterOptionsType]?.isDynamic;
+    if (isDynamic) {
+      const values = ensureArray(value);
+      if (values.length > 0) dynamicFilters.push({ key, values });
+    } else {
+      staticFilterValues[key] = value;
+    }
+  }
+
+  if (dynamicFilters.length > 0) {
+    return { ...staticFilterValues, dynamicFilters };
+  }
+  return staticFilterValues;
 };
